moved plugins load outside the CommandHandler
